Prime Numbers Loop
Have the user input two numbers, a low and high number. Find all of the prime numbers between the low and high, and print them.
Count how many prime numbers were found and if that number is prime, print "SUCCESS" an amount equal to the number of primes found, then tell the user the program has ended.
If the number was not prime tell the user that they failed and need to try again. Loop the program until the user gets a successful prime number.