Read in from another file brownie program
Read in from a file and tell how many numbers there are, and what each number is. Tell the user what the smallest and largest numbers are, as well as the average of all numbers and the standard deviation. Find the ((average-smallest)/standard deviation) and tell the user if the batch should be rejected.